test: trabajando en test de locust. Hecho test de locust para probar registro y actualizacion
